@@ -1,14 +1,14 @@
 apply plugin: 'com.android.library'
 
 android {
-    compileSdkVersion 27
-    buildToolsVersion "27.0.3"
+    compileSdkVersion 29
+    buildToolsVersion "29.0.2"
 
     defaultConfig {
-        minSdkVersion 14
-        targetSdkVersion 27
-        versionCode 28
-        versionName "2.7"
+        minSdkVersion 16
+        targetSdkVersion 29
+        versionCode 29
+        versionName "2.9"
     }
 
     buildTypes {
@@ -25,8 +25,8 @@ android {
 
 dependencies {
     api fileTree(dir: 'libs', include: ['*.jar'])
-    api 'com.android.support:appcompat-v7:27.1.1'
-    api 'com.android.support:design:27.1.1'
+    api 'androidx.appcompat:appcompat:1.0.0'
+    api 'com.google.android.material:material:1.0.0'
     api 'com.squareup.okhttp3:okhttp:3.10.0'
     api 'org.jsoup:jsoup:1.11.3'
     api 'com.mindorks.android:prdownloader:0.6.0'
